QtTest: Fix format error in _q_compareSequence()
Commit 63295f43e70955a194602a7acf39013f124e14b7 changed the size:
placeholders to %lld, but wrapped only one of the arguments in
qlonglong. This causes 32-bit builds to read past the end of the
expectedSize variable.

Fix by wrapping the other, too.

Pick-to: 6.5 6.2 5.15
(cherry picked from commit 57eff5b092dbd692594125edc70d3efd898660ea)
(cherry picked from commit f92cab0bee90e4349550676f0f4bdf240f4d3669)

+++ src/testlib/qtest.h
@@ -435,7 +435,7 @@ bool _q_compareSequence(ActualIterator actualIt, ActualIterator actualEnd,
                       "   Actual   (%s) size: %lld\n"
                       "   Expected (%s) size: %lld",
                       actual, qlonglong(actualSize),
-                  expected, expectedSize);
+                      expected, qlonglong(expectedSize));
     }
 
     for (auto expectedIt = expectedBegin; isOk && expectedIt < expectedEnd; ++actualIt, ++expectedIt) {
